libvirtd_start.py: Added autostart of inactive storage pools to the list of accepted errors.
This patch fixes the issue of libvirtd service failing due to autostart of inactive storage pools,
which is irrelevant to the current test-cases of iptables/firewalld start and stop along with libvirtd.

Also modified the testcase to SKIP where Selinux needs to be in enforcing mode, as its not supported
on certain distributions like Ubuntu.

Signed-off-by: Santwana Samantray <santwana@linux.vnet.ibm.com>